pubsubpull/requirements.txt
Flask==0.11.1
gunicorn==19.6.0
google-cloud-pubsub==0.29.0
google-api-python-client
oauth2client==3
pandas
google-cloud-logging
pandas-gbq
utils